Habana device manager.
