+++ lib/Backends/DeviceManagers.cpp
@@ -17,6 +17,7 @@
 #include "glow/Backends/DeviceManager.h"
 #include "glow/Backends/DummyDeviceManager.h"
 #include "glow/Support/Register.h"
+
 #include "llvm/Support/Casting.h"
 #include "llvm/Support/raw_ostream.h"
