+++ lib/Backends/Habana/HabanaUtils.cpp
@@ -17,6 +17,7 @@
 #include "HabanaUtils.h"
 
 namespace glow {
+
 const char *statusStr(synStatus status) {
   static const char *strs[] = {
       "Success",
@@ -48,4 +49,39 @@ const char *statusStr(synStatus status) {
   }
   return strs[status];
 }
+
+bool allows64To32Downcast(const Node *src, const Node *dst) {
+  // If N is used as the indices or weights of a sparse lookup, it is safe to
+  // access a partial tensor.
+  if (auto *SLS =
+          llvm::dyn_cast<FusedRowwiseQuantizedSparseLengthsWeightedSumNode>(
+              dst)) {
+    return src == SLS->getIndices();
+  } else if (auto *SLS =
+                 llvm::dyn_cast<FusedRowwiseQuantizedSparseLengthsSumNode>(
+                     dst)) {
+    return src == SLS->getIndices();
+  } else if (auto *SLS = llvm::dyn_cast<SparseLengthsWeightedSumNode>(dst)) {
+    return src == SLS->getIndices();
+  } else if (auto *SLS = llvm::dyn_cast<SparseLengthsSumNode>(dst)) {
+    return src == SLS->getIndices();
+  }
+  return false;
+}
+
+bool allows64To32Downcast(const Storage *V, const Function *F) {
+  if (V->getElementType() != ElemKind::Int64ITy) {
+    return false;
+  }
+  for (auto const &U : V->getUsers()) {
+    if (U.getUser()->getParent() != F) {
+      continue;
+    }
+    if (!allows64To32Downcast(*U.get(), U.getUser())) {
+      return false;
+    }
+  }
+  return true;
+}
+
 } // namespace glow
